requirements.txt
pytest
django
djangorestframework
pytest-django